feat: Adicionar confirmação antes de excluir funcionário e botões de navegação para melhorar o fluxo do usuário

+++ src/pages/Employees.tsx
@@ -1,5 +1,7 @@
 import { useEffect, useState } from "react"
 import type { Employee } from "../types/employee"
+import { useNavigate } from "react-router-dom"
+
 
 function Employees() {
     const [employees, setEmployees] = useState<Employee[]>([])
@@ -14,18 +16,28 @@ function Employees() {
 
 
     function handleDelete(indexToDelete: number) {
+        const confirmDelete = window.confirm(
+            "Tem certeza que deseja excluir este funcionário?"
+        )
+
+        if (!confirmDelete) return
+
         const updatedEmployees = employees.filter(
             (_, index) => index !== indexToDelete
         )
 
         setEmployees(updatedEmployees)
-
         localStorage.setItem("employees", JSON.stringify(updatedEmployees))
     }
 
+    const navigate = useNavigate()
 
     return (
         <div>
+            <button onClick={() => navigate("/")}>
+                ← Voltar para Home
+            </button>
+
             <h1>Funcionários Cadastrados</h1>
 
             {employees.length === 0 ? (
